tokio_signal

Module unix

Unix-specific types for signal handling.

This module is only defined on Unix platforms and contains the primary Signal type for receiving notifications of signals.

Re-exports

pub extern crate libc;

Structs

Signal

An implementation of Stream for receiving a particular type of signal.

Constants

SIGALRM

SIGHUP

SIGINT

SIGPIPE

SIGQUIT

SIGTERM

SIGTRAP

SIGUSR1

SIGUSR2